pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ocrscout"
version = "0.1.0"
description = "Scout frontier OCR models on your data, your hardware, your terms."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "Sebastian Majstorovic", email = "storytracer@gmail.com" },
]
keywords = ["ocr", "vlm", "docling", "evaluation", "benchmark"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
]
dependencies = [
    "docling-core>=2.50.0",
    "huggingface-hub>=0.25.0",
    "datasets>=3.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyarrow>=17.0.0",
]

[project.optional-dependencies]
pdf = ["ocrmypdf>=16.0.0", "pikepdf>=9.0.0"]
alto = ["lxml>=5.0.0"]
iiif = ["httpx>=0.27.0"]
docling = ["docling"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[project.scripts]
ocrscout = "ocrscout.cli:main"

[project.urls]
Homepage = "https://github.com/storytracer/ocrscout"
Issues = "https://github.com/storytracer/ocrscout/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ocrscout"]

[tool.hatch.build.targets.wheel.force-include]
"src/ocrscout/profiles" = "ocrscout/profiles"

[tool.hatch.build.targets.sdist]
include = ["src/ocrscout", "tests", "README.md", "CLAUDE.md", "LICENSE"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]  # typer.Option in defaults is the canonical typer idiom

[tool.mypy]
python_version = "3.11"
strict_optional = true
warn_unused_ignores = true
